+++ src/UserSegment.js
@@ -1,0 +1,12 @@
+import AppLovinMAX from "./index.js";
+
+let UserSegment = {
+
+  set name(value) {
+    AppLovinMAX.setUserSegment(value);
+  },
+};
+
+export {
+  UserSegment,
+};