+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## x.x.x
     * Add support for waterfall.
+    * Add support for setting user segment via `AppLovinMAX.userSegment.name`.
 ## 3.1.0
     * Add support for data passing.
     * Add API for Android to provide a built-in consent flow that sets the user consent flag.